+++ python/kserve/kserve/handlers/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2021 The KServe Authors.
+# Copyright 2022 The KServe Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,9 +13,7 @@
 # limitations under the License.
 
 from .base import NotFoundHandler  # noqa # pylint: disable=unused-import
-from .liveness import LivenessHandler  # noqa # pylint: disable=unused-import
-from .health import HealthHandler  # noqa # pylint: disable=unused-import
-from .list import ListHandler  # noqa # pylint: disable=unused-import
-from .load import LoadHandler, UnloadHandler  # noqa # pylint: disable=unused-import
+from .health import LivenessHandler, HealthHandler  # noqa # pylint: disable=unused-import
+from .model_management import LoadHandler, UnloadHandler, ListHandler  # noqa # pylint: disable=unused-import
 from .explain import ExplainHandler  # noqa # pylint: disable=unused-import
 from .predict import PredictHandler  # noqa # pylint: disable=unused-import
